rdctl: refactor `rdctl shell`
This moves the bulk of the code into a separate package so that it can be
reused elsewhere.  It returns the command without executing so that it can
be further customized by the caller.
